Connection_id : i32

+++ src/abstractions.rs
@@ -28,7 +28,7 @@ pub trait MyServiceBusSubscriberClient {
         topic_id: &str,
         queue_id: &str,
         confirmation_id: i64,
-        connection_id: i64,
+        connection_id: i32,
         delivered: bool,
     );
 
@@ -37,7 +37,7 @@ pub trait MyServiceBusSubscriberClient {
         topic_id: &str,
         queue_id: &str,
         confirmation_id: i64,
-        connection_id: i64,
+        connection_id: i32,
         ok_messages: crate::queue_with_intervals::QueueWithIntervals,
     );
 }
@@ -52,6 +52,6 @@ pub trait MyServiceBusSubscriberClientCallback {
         &self,
         messages_to_deliver: Vec<MySbMessage>,
         confirmation_id: i64,
-        connection_id: i64,
+        connection_id: i32,
     );
 }
